backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
firebase-admin>=6.5.0
requests>=2.31.0
transformers>=4.39.0
torch>=2.2.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
